backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0
playwright-stealth>=1.0.0

# Database
chromadb>=0.4.22

# Agent infrastructure
tiktoken>=0.5.0
sse-starlette>=1.6.0

# Utilities
tqdm>=4.66.0
